Add tag service and update repository file options
Tags is now a separate service, following the Gitlab API structure.
