+++ .travis.yml
@@ -1,5 +1,5 @@
 language: go
 
 go:
-  - 1.5.4
   - 1.6.3
+  - 1.7.1
