+++ gitlab.go
@@ -133,6 +133,7 @@ type Client struct {
 	Session         *SessionService
 	Settings        *SettingsService
 	SystemHooks     *SystemHooksService
+	Tags            *TagsService
 	Users           *UsersService
 }
 
@@ -189,6 +190,7 @@ func newClient(httpClient *http.Client, tokenType tokenType, token string) *Clie
 	c.Session = &SessionService{client: c}
 	c.Settings = &SettingsService{client: c}
 	c.SystemHooks = &SystemHooksService{client: c}
+	c.Tags = &TagsService{client: c}
 	c.Users = &UsersService{client: c}
 
 	return c
